Configued clangd not *never* auto-insert include
The convention in most project I work with is to use angle brackets
for include directives, e.g. #include <path/to/lib/feature.h>.
However, clangd when auto inserting include directives uses quotes
and that is annoying. Quick investigation seems to indicate that this is
not configurable in clangd. So better to leave it out.